make: changes to support dpdk submodule with a relative path
Make the dpdk submodule in spdk the default dpdk library. The dpdk submodule
is located in the dpdk director at the top level of the spdk tree.

To use the dpdk submodule at Broadocom:

  1. git clone https://{username}@gitlab.com/spdk-fc-nvme/spdk.git
  2. cd spdk
  3. git config submodule.dpdk.url https://{username}@gitlab.com/spdk-fc-nvme/dpdk.git
  4. git submodule init
  5. git submodule update

To use the dpdk submodule at NetApp:

  1. git clone https://bitbucket.eng.netapp.com/scm/nvmf-tgt-bb/spdk.git
  2. cd spdk
  3. git config submodule.dpdk.url https://bitbucket.eng.netapp.com/scm/nvmf-tgt-bb/dpdk.git
  4. git submodule init
  5. git submodule update

Following this "make", "make clean" and "make CC=clang CXX=clang++" will automatically
build, clean and rebuild dpdk at the same time that it builds spdk. This leverages
the integrated support for dpdk in the spdk Makefiles.

To overide this behavior to use the old method, simply say:

 make CONFIG_DPDK_DIR=../dpdk/build

This will ignore the dpdk submodule in spdk, even it's it there.