nvmf_fc: ABTS response not being sent out.
There is a bug in the ABTS optimization logic that decides on
sending the ABTS to the different HWQPs. The bug results in the
ABTS frames being only sent if they were received on the first physical
port. For subsequent ports (hwqp_id greater than 16), this logic
does not work.

Fixed the issue.