nvmf: make Asynchronous Event Configuration feature programmable
 - Add support for progaming Feature Identifier 0Bh
 - Prevents the setting of unsupported async events
 - Spdk virtal controller currently only supports
   Namespace Attribute Notices (0x100)

5.21.1.11 Asynchronous Event Configuration (Feature Identifier 0Bh)

This Feature controls the events that trigger an asynchronous event
notification to the host. This Feature may be used to disable reporting events
in the case of a persistent condition (refer to section 5.2). If the condition
for an event is true when the corresponding notice is enabled, then an event
is sent to the host. The attributes are indicated in Command Dword 11.
If a Get Features command is submitted for this Feature, the attributes
specified in Figure 148 are returned in Dword 0 of the completion queue entry
for that command.

Figure 142: Asynchronous Event Configuration  Command Dword 11

Bit           Description
10            Telemetry Log Notices
09            Firmware Activation Notices
08            Namespace Attribute Notices
07:00         SMART/Health Critical Warnings