nvmf_fc: Common XRI ring between SCSI and NVMeoF
These changes bring in a common XRI ring betwen the SCSI and the NVMeoF
protocols. The XRI ring, its associated SGLs and all the send-frame XRIs
are allocated in a common place in the low level driver. The pointer to
the XRI ring and the send-frame XRIs is passed via the init-args to the
SPDK transport (via fc-adm-api).