nvmf_fc: Value of sc in nvme-status not being reset
This was a bug that I caught while building the UT infrastructure.

The "sc" value is not being set to 0 when a new request is instantiated.
So, if a previous request had an "error" sc value, then, the it
is possible that this value gets sent across even if the output of the
command was succesful. Basically, we don't set the sc value in every
success case.

The issue was found after running a negative test for identify controller
that set the value of sc to an error.
The next test was a positive test for testing set-property message and
that showed this problem.