nvmf_fc: cdw0 not initialized in resp payload
This can potentially cause many issues. But, the main one that I
found was that having a non-zero value in cdw0 meant that we sent
a ERSP for every READ command. This has an overall performance
impact. The "ar" bit was NOT being set for a READ message whenever the
cdw0 was incorrectly set to non-zero.

After discussion with Brcm, we will memset the cmpl field to 0 to keep
it future safe.