nvmf_fc: RAS improvements.
This change adds a few improvements in the area of tracking pending commands.
1. The LS pending commands are tracked with an entry and exit timestamp. The
average time spent on the pending queue, the number of LS commands processed
and the highest time spent on the pending queue is tracked.
2. On the IO side there was a small bug where there was the potential for the
timestamp on the IO on the pending queue being reset. Fixed that issue.
3. Track total number of commands on a connection pending queue. Offline counter
tools can scrape this data whenever needed to understand typical command fan-in
pressure on a HWQP.
4. Add a counter to track total IO count on a HWQP. Fix the connection field
related to queue depth so that it can be looked at to understand the IO queue
depth from a per connection basis as well.
5. Based on code review added a counter to track number of LS commands in the
pending queue.