nmvf_fc: Pending FC-LS command cleanup and connection fairness
1)
Today, a LS request gets parked when we dont get a XRI.
If a LS request gets parked in the pending queue and we get a
IT delete (or a Nport delete), we clean up all outstanding requests.
We also clean up any existing associations and then delete the
Rport/Nport. But we dont look into the pending queue. So, eventually
the LS request may get "processed" from the pending queue.
Now, there are a couple of checks there to see that the Nport/Rport
have not been deleted when we start working on the LS request.
But, we discovered an issue where the host restablished the
rport after deletion.
The new incarnation of the Rport had the same SID/DID as the original
one. So, the LS processing started. But, in this case, the Rport
pointer was invalid since the original memory had been freed.

The fix is to drain the LS pending queue when there are Rport deletions.

2)
The other issue is fairness is looking to be a concern in how we pull
items out of the connection pending list for IOs.
If we have a system with more IOs slated than the number of XRIs then,
it is possible that we are always pulling out the IOs from the pending
list in the same order. i.e. some connections can get starved for ever.

The fix is to provide some sort of fairness on the connections in a HWQP.
This is done by pulling the connection from the head of the list when
being processed and placing it at the tail.

Updates with code review comments