nvmf_fc: Fairness for pending requests over new ones on a HWQP
In our current logic, we have situations where we allow new commands
from the RQ to be handled before the pending ones are handled on a
HWQP. This is typically seen when we have overloaded systems with XRI
pressure. In such cases, one HWQP may be freeing up XRIs by finishing
cq_wq. At the same time another HWQP may process RQ buffers.
It will find XRIs and then process these commands even if there were
earlier requests stuck in the pending queue.

We had a test where about 16K commands were being pumped into one
target port using 8 hosts.
Looking into the core, I see commands in the pending queue for
8-12 seconds. The system had enough CPU (40 core AFF box)
to do the work so it looks like fairness was the key issue here.

I have a patch attached here for this issue.

In the patch I have also fixed a very critical bug that Naresh found regarding
deletion of ls pending requests. This was a regression introduced from my side
where we end up deleting all LS pending requests instead of only the required
ones for the Nport/Rport. It probably does not matter in most cases since we
come down this path for f/w dumps or resets (when all Nports are gone).
But, we should fix this as well.

This is going to be a VR candidate.